+++ napari/layers/_layer_actions.py
@@ -14,6 +14,7 @@ from napari.layers import Image, Labels, Layer
 from napari.layers._source import layer_source
 from napari.layers.utils import stack_utils
 from napari.layers.utils._link_layers import get_linked_layers
+from napari.utils.migrations import _DeprecatingDict
 from napari.utils.translations import trans
 
 if TYPE_CHECKING:
@@ -26,7 +27,7 @@ def _duplicate_layer(ll: LayerList, *, name: str = '') -> None:
     for lay in list(ll.selection):
         data, state, type_str = lay.as_layer_data_tuple()
         state['name'] = trans._('{name} copy', name=lay.name)
-        # TODO: fix typing here.
+        state = cast(_DeprecatingDict, state)
         for deprecated in state.deprecations:
             del state[deprecated]
         with layer_source(parent=lay):
